entity: enable caching a pipeline with an entity
When painting entities we want to be able generate a pipeline based on
the components associated with the entity and potentially some other,
editor specific state. This provides a simple mechanism for editors to
associate a generated pipeline with an entity to avoid repeated
recreation of the pipeline each frame.