editor: traverse the scenegraph for picking
Instead of having to maintain a separate list of "pickables" we now
traverse the scenegraph looking for entities with pickable geometry.

This also changes the code to use the framebuffer's matrix stack so
we can accumulate entity transforms as we traverse the graph so we
can avoid repeatedly accumulating the transforms of common ancestors.