Handle more of the view transform via entities
To break the components of the camera view transform down so they can
easily be manipulated to handle the keyboard+mouse navigation we are
aiming to represent the camera view as a concatenation of entity
transforms.

The current model is as if the camera were mounted on a physical
armature that can pivot and change length. We should be able to handle
zooming by changing the length of the armature, translation by moving
the whole armature, and rotations around the origin by pivoting the
armature.