editor: Adds initial undo-redo support
This adds some initial data types and apis to be able to log all
operations that modify the current UI so we can undo or redo them. The
aim is that all operations should be invertible and that we only ever
append to the log such that it also supports undoing undos.