Add the ability for paintable objects to paint in multiple layers
Paintable objects can now paint in multiple layers so that parts of
their interface can appear above all other normal interface
components. For example, a combo box would want its drop down menu to
appear above any other buttons it is next to. However it is
conceptually part of the same node in the graph and will have the same
transform so it would be awkward to have to add a separate node in the
graph for it.

The RigPaintContext now contains a list of objects that need to be
painted again. A paintable object can add itself to this list with a
convenience function while its implementation of the paint method is
called. After initially walking the graph, the painting algorithm is
expected to repeatedly walk this list until it becomes empty. Each
entry in the list also gets a copy of the modelview matrix that was
current when the node was added so that it doesn't need to be
recalculated for the subsequent paints. There is also a layer_number
member of RigPaintContext so that the paintables can detect which
layer is currently being painted. The object can add itself to the
list multiple times to get painted in multiple layers. There is a
convenience function in rig-paintable to implement iterating the graph
and the layers.