Add RigInspector
RigInspector represents a control which can edit all of the properties
on a given object. The properties are introspected and appropriate
controls for them are created at run-time.

The inspector is constructed with a callback to actually update the
property when a control is used. The callback is passed the source and
target properties so that it can read the value before and after and
copy it across.