inspector: avoid binding conflicts
Since something may have already defined a binding for a property we
check first before assuming its ok to bind it to an inspector widget.
For example the editor handles animatable properties via bindings
and the editor needs to be kept in the loop about property changes
so it can track key frame animation values.