Use a callback list pattern for objects that need a callback
Previously there was a pattern of having a single callback function
when an object needs to notify its owner about some change. This patch
changes everything that was doing that to use a list of callbacks
instead. That way multiple listeners could be registered.

There is a RigClosure type with some convenience functions and a macro
to manage a list of callbacks. The function to add a callback is
expected to directly return a RigClosure pointer. The caller can use
this pointer to disconnect the callback using the general purpose
rig_closure_disconnect() function. That way each callback point
doesn't have to have a separate function to disconnect the callback.

The callbacks are invoked directly by the object so that there doesn't
need to be any fiddly marshalling like there would be with GSignals.
However there is also a convenience macro which takes variadic
arguments and will call each callback with the given arguments.