editor: Use SplitView widgit to layout ui
Instead of rigidly defining the layout of the ui with a set RigTransform
nodes and an allocate() function that's called when the window is
resized we now use the RigSplitView to divide the window vertically and
horizontally to position ui elements.

There are still various limitations and problems with the split-view
layouting at this point but at least now we have the basic ability to
resize internal elements of the ui.

In addition to using RigSplitView we also scatter a few bevels around
so the ui doesn't look quite so plane.