shell: skip cameras if input handled in scene traversal
Each input camera can have an associated scenegraph of nodes that we
pick against and if successful we handle the event. Because
rig_graphable_traverse didn't previously return any status information
we couldn't tell if the event had been handled though and so we would
always continue to consider other input cameras and their corresponding
scene graphs, potentially resulting in an event being handled multiple
times. Since we now know when a traversal is cut short we can determine
when the event was handled and so we now avoid considering further input
cameras once the event has been handled.