Add a boolean to mark whether each property is animatable or animated
Each RigPropertySpec now has a boolean to mark whether it should be
animatable. The user should also be able to select whether an
individual property is animated so there is now also a boolean on each
RigProperty for that. This patch also sets the animatable field in the
property specs for RigEntity and RigCamera.

There is now a RigClosureList in the RigPropertyContext that will get
invoked whenever the animated state of any property is changed. This
will be used by the editor to update the timeline whenever the state
is changed. The animated flag should be toggled by calling
rig_property_set_animated.