inspector: Add a separate widget for editing each property
The controls for editing an individual property are now handled by a
separate widget called a RigPropInspector. The RigPropInspector can
have more than one control for each property. Keeping this
encapsulated in one widget means the layouting code for the
RigInspector doesn't need to change. The RigPropInspector just lays
out its controls in a horizontal line.

Currently the only extra control per property is a toggle button to
toggle whether the property is animated. This is only added if the
spec for the property is marked as animatable.