toggle: Add properties to set the tick glyph and color
It would be nice to be able to change the appearance of the toggle
button somewhat when using it to set whether a property is animated so
that it can be distuinguished from boolean properties. This patch adds
some properties to change the tick glyph and the color so that the
tick can be changed to a red dot for the recording control.