Fix get_vertex_data implementation for RigDiamond and RigMeshRenderer
The implementations for these two components were incorrectly checking
whether the n_vertices out variable was passed as a NULL pointer or
not. Depending on whatever garbage was left in the memory it would
either write the number of vertices or leave it as zero. This made
picking stop working in the editor.