transition: Add a 'constant' value to each property
Previously a RigTransition contained a GList which had a path for each
property that it knew about. This replaces the GList with a hash table
containing a separate struct for each property. The struct contains
both a constant value and a path. The constant value is intended to be
used whenever the property is not animated. It is stored in a
RigBoxed.

The RigPath no longer knows which property it is attached to and is
instead just constructed with a type. It no longer binds to a progress
property and instead the transition directly updates the animated
properties whenever its progress changes.

The loading and saving code has been adapted to handle storing the
constant value. Instead of a <transition> tag containing a list of
<path>s as direct children, it now contains <property> tags instead.
The property tag contains all of the attributes that were previously
attached to the path tag and it additionally has an 'animated'
attribute. The property tag now contains a <constant> child to store
the constant value and an optional path tag to store the path. That
looks something like this:

<transition id="0">
  <property entity="6" name="position" animated="no">
    <path>
      <node t="0.000000" value="(332.253418, 637.272949, 0.000000)" />
    </path>
    <constant value="(0.000000, 0.000000, 0.000000)" />
  </property>
</transition>

The code to convert to and from string representations of values have
been split into a series of functions so that it can be shared between
the code to load a node value and to load a boxed value. For
completeness this patch also adds code and node types for all of the
missing property types.