transition: Add a function to update the entity's property
This adds rig_transition_update_property which updates the property on
an entity depending on the current settings within the transition. If
the property is animated then it will interpolate the current value
based on the path and the current transition progress, otherwise it
will read the value from the property's constant value.