editor: Give the light a cube handle
This associates a cube mesh with the light so that it can be selected
and manipulated in the scene. The mesh isn't directly set on the light
because we need to scale the mesh to make it visible but don't want to
scale the light since that would affect the rendering of shadow maps. A
light_handle entity is created as a child of the light which we
associate the map with and as a special case when we pick to select an
entity we check to see if the light_handle has been selected then we
instead select the light. We can potentially use a similar trick for
associating a handle with a camera.