editor: draw cameras without rig_entity_draw
The editor now draws cameras using rig_camera_create_frustum_primitive
to create a CoglPrimitive representing the camera's frustum and manually
draws that on a framebuffer. The guard inside rig_camera_draw() to
determine when to draw a wireframe or when to actually prepare to draw
the camera wasn't good enough so it makes sense to let the editor decide
exactly what it wants to do.