Remove cogl-primitives.h and CoglTextureVertex
This removes the cogl-primitives.h header. The implementations for all
of the functions that it declared have already been removed in
7546ee19.

The CoglTextureVertex type has also been removed because it was only
used for cogl_polygon which has already been removed.

This patch also fixes some mentions of cogl_rectangle() in the
documentation and replaces them with
cogl_framebuffer_draw_rectangle().

Reviewed-by: Robert Bragg <robert@linux.intel.com>