Make RigBindingCallback also pass the source property
If a property binding depends on multiple properties it's useful for
the callback to be passed the property which caused the dependant
property to change. That way it can read from only the single property
that has changed.

This patches changes the signature of the callback and also all of the
implementations.