vec3-slider: Don't emit inconsistent state when the value is set
When the value is directly set on a vec3 slider it would previously
just update the values in the child sliders and let the dirty
notification on that cause a dirty notification from the vec3 slider
to be emitted. However because the dirty notifications are synchronous
updating the first slider would immediately cause the vec3 to
advertise the new vec3 value with only the first component changed
which is an inconsistent state.

To fix this the setter for the value property now immediately updates
its cached value with the passed in value before updating the child
number sliders. The property binding callback for the number slider
now only reads the value for the number slider that has actually
changed instead of always reading all three of them. That way the vec3
slider will always have a consistent cached value property.