Add functions to reload the inspector properties
This adds rig_prop_inspector_reload_property. This causes the
RigPropInspector to reread the value that it is bound to and copy it
into the control used for editing.

The rig_inspector_reload_property works similarly but takes a pointer
to the property to reload. It then causes the corresponding
RigPropInspector to reload its value.