editor: Reload inspector's position property when moving an entity
Whenever an entity is dragged it now calls
rig_inspector_reload_property with the position property so that the
new values will be reflected in the inspector.

This should eventually be replaced with a more generic mechanism for
the editor to update any property. That should be able to log the
property change in the undo history and either update the transition's
constant value or add a node to the path depending on whether it is
animated. The call to rig_inspector_reload_property should be hooked
into this single function.