Rename util code to 'rut' and editor/runtime to 'rig'
Since the editor can be conditionally disabled leaving just a runtime it
doesn't make sense to have an "editor" directory name for rig. Also it
was starting to get awkward to avoid namespace clashes between symbols
that are internal to the editor/runtime and utility api symbols.

This changes the namespace of utility apis to "rut" (Rig Utilities)
which is intentionally three characters long to ease movement of code
from rig/ to rut/