Remove the deprecated feature flags
The old system for checking features depended on a bitmask in an
integer which meant we could only have up to 32 features. This was
deprecated in 426c8b8f41. The API for using the legacy feature mask
has already been removed so we might as well also remove the internal
code to maintain the bitmask.

Reviewed-by: Robert Bragg <robert@linux.intel.com>